dev_requirements.txt
python-dotenv
supervisely==6.73.259
# formatter
black

# sorter
isort
